Split operations

+++ atcoder-problems-predictor/rate_predictor.py
@@ -13,6 +13,8 @@ import pickle
 
 COLUMN_RATING = "Rating"
 COLUMN_PREDICT = "Predict"
+PROBLEM_SET_JSON_NAME = "./problem_set.json"
+MODEL_DUMP_NAME = "./save_xgb_predicted_rating"
 
 
 def get_submissions(users: List[str], conn, table_name: str) -> List[Tuple[str, str, str, int]]:
@@ -148,26 +150,27 @@ def main(filepath: str, command: str):
         config = json.load(f)
     conn = psycopg2.connect(config["db"])
 
-    model = xgb.XGBRegressor()
-    problem_set = set()
+    if command is "train":
+        model = xgb.XGBRegressor()
+        problem_set = set()
 
-    train_model(model, problem_set, conn)
-    pickle.dump(model, open("./save_xgb_predicted_rating", "wb"))
-    json.dump(list(problem_set), open("./problem_set.json", "w"))
+        train_model(model, problem_set, conn)
+        pickle.dump(model, open(MODEL_DUMP_NAME, "wb"))
+        json.dump(list(problem_set), open(PROBLEM_SET_JSON_NAME, "w"))
+    else:
+        loaded_model = pickle.load(open(MODEL_DUMP_NAME, "rb"))
+        loaded_set = set(json.load(open(PROBLEM_SET_JSON_NAME, "r")))
 
-    loaded_model = pickle.load(open("./save_xgb_predicted_rating", "rb"))
-    loaded_set = set(json.load(open("./problem_set.json", "r")))
-
-    predicted_result = predict(loaded_model, loaded_set, conn)
-    with conn.cursor() as cursor:
-        cursor.execute("DELETE FROM predicted_rating")
-        for user_id, rate in predicted_result.items():
-            query = """
-                    INSERT INTO predicted_rating (user_id, rating)
-                    VALUES (%s, %s)
-                    """
-            cursor.execute(query, (user_id, rate))
-            conn.commit()
+        predicted_result = predict(loaded_model, loaded_set, conn)
+        with conn.cursor() as cursor:
+            cursor.execute("DELETE FROM predicted_rating")
+            for user_id, rate in predicted_result.items():
+                query = """
+                        INSERT INTO predicted_rating (user_id, rating)
+                        VALUES (%s, %s)
+                        """
+                cursor.execute(query, (user_id, rate))
+                conn.commit()
 
 
 if __name__ == '__main__':
